[spark] Consolidate scan builders into Append/Upsert with SplitPlanner
Consolidates the four scan builders (FlussLakeAppendScanBuilder /
FlussAppendScanBuilder / FlussLakeUpsertScanBuilder / FlussUpsertScanBuilder)
into two (FlussAppendScanBuilder, FlussUpsertScanBuilder), routing lake-union
vs log-only reads through a single SplitPlanner probed at scan build time.

Highlights:

- Introduce SplitPlanner with AppendPlanner / UpsertPlanner. Each planner
  probes the readable lake snapshot exactly once at construction and
  dispatches inside plan() to either the lake-union branch (lake splits
  unioned with a Fluss log-tail / kv+log-tail) or the log-only branch
  (pure Fluss read). No runtime fallback path — presence/absence of the
  lake snapshot is decided deterministically at construction.

- Remove FlussLakeAppendBatch, FlussLakeUpsertBatch, FlussLakeBatch. The
  planner-backed FlussAppendBatch / FlussUpsertBatch dispatch the reader
  factory based on planner.hasLakeSnapshot, and own the planner's
  Fluss client connection / admin lifecycle via close().

- Batch scans no longer consume SCAN_START_UP_MODE. Batch semantics are
  "the full table" regardless of the user-facing startup mode. This keeps
  append and upsert planners symmetric (KV snapshot has no partial-read
  semantics) and structurally avoids the empty offset-range case that
  otherwise trips the reader-side guard. Time-range batch reads should
  be expressed via predicate pushdown, not startup mode. AppendPlanner
  uses OffsetsInitializer.full() rather than earliest() so start offsets
  resolve to concrete numeric values usable by the split-by-max-records
  logic.

- Merge FlussSupportsPushDownV2Filters and FlussLakeSupportsPushDownV2Filters.
  Pushdown decisions are schema-level (partition predicate extraction plus
  ARROW/non-PK data-predicate acceptance); further pushdown to LakeSource
  happens inside the planner and is transparent to Spark.

- SparkTable is no longer lake-aware. newScanBuilder constructs a plain
  Append/Upsert ScanBuilder; the planner probes the snapshot internally.

- Tests: SparkPrimaryKeyTableReadTest asserts SCAN_START_UP_MODE is ignored
  for batch reads; SparkLakeTableReadTestBase helper no longer filters by
  planner type.
