+++ fluss-spark/fluss-spark-common/src/main/scala/org/apache/fluss/spark/read/FlussScan.scala
@@ -21,7 +21,6 @@ import org.apache.fluss.config.Configuration
 import org.apache.fluss.metadata.{TableInfo, TablePath}
 import org.apache.fluss.predicate.{Predicate => FlussPredicate}
 import org.apache.fluss.spark.SparkConversions
-import org.apache.fluss.spark.read.lake.{FlussLakeAppendBatch, FlussLakeUpsertBatch}
 
 import org.apache.spark.sql.connector.expressions.filter.Predicate
 import org.apache.spark.sql.connector.metric.CustomMetric
@@ -70,7 +69,11 @@ trait FlussScan extends Scan {
     Array(FlussNumRowsReadMetric())
 }
 
-/** Fluss Append Scan. */
+/**
+ * Fluss Append (log-table) scan. Whether the underlying batch reads from Fluss only or unions
+ * Fluss with a lake snapshot is determined by the [[AppendSplitPlanner]] instance passed in from
+ * the ScanBuilder. Description reflects the planner category.
+ */
 case class FlussAppendScan(
     tablePath: TablePath,
     tableInfo: TableInfo,
@@ -80,10 +83,12 @@ case class FlussAppendScan(
     override val pushedSparkPredicates: Seq[Predicate],
     override val limit: Option[Int],
     options: CaseInsensitiveStringMap,
-    flussConfig: Configuration)
+    flussConfig: Configuration,
+    planner: AppendSplitPlanner)
   extends FlussScan {
 
-  override protected val scanType: String = "Append"
+  override protected val scanType: String =
+    if (planner.hasLakeSnapshot) "LakeAppend" else "Append"
 
   override def toBatch: Batch = {
     new FlussAppendBatch(
@@ -91,10 +96,10 @@ case class FlussAppendScan(
       tableInfo,
       readSchema,
       pushedPredicate,
-      partitionPredicate,
       limit,
       options,
-      flussConfig)
+      flussConfig,
+      planner)
   }
 
   override def toMicroBatchStream(checkpointLocation: String): MicroBatchStream = {
@@ -108,8 +113,12 @@ case class FlussAppendScan(
   }
 }
 
-/** Fluss Lake Append Scan. */
-case class FlussLakeAppendScan(
+/**
+ * Fluss Upsert (primary-key table) scan. Whether the underlying batch reads from Fluss only or
+ * unions Fluss with a lake snapshot is determined by the [[UpsertSplitPlanner]] instance passed
+ * in from the ScanBuilder.
+ */
+case class FlussUpsertScan(
     tablePath: TablePath,
     tableInfo: TableInfo,
     requiredSchema: Option[StructType],
@@ -118,94 +127,23 @@ case class FlussLakeAppendScan(
     override val pushedSparkPredicates: Seq[Predicate],
     override val limit: Option[Int],
     options: CaseInsensitiveStringMap,
-    flussConfig: Configuration)
+    flussConfig: Configuration,
+    planner: UpsertSplitPlanner)
   extends FlussScan {
 
-  override protected val scanType: String = "LakeAppend"
-
-  override def toBatch: Batch = {
-    new FlussLakeAppendBatch(
-      tablePath,
-      tableInfo,
-      readSchema,
-      pushedPredicate,
-      partitionPredicate,
-      limit,
-      options,
-      flussConfig)
-  }
-
-  override def toMicroBatchStream(checkpointLocation: String): MicroBatchStream = {
-    new FlussAppendMicroBatchStream(
-      tablePath,
-      tableInfo,
-      readSchema,
-      options,
-      flussConfig,
-      checkpointLocation)
-  }
-}
-
-/** Fluss Upsert Scan. */
-case class FlussUpsertScan(
-    tablePath: TablePath,
-    tableInfo: TableInfo,
-    requiredSchema: Option[StructType],
-    override val partitionPredicate: Option[FlussPredicate],
-    override val limit: Option[Int],
-    options: CaseInsensitiveStringMap,
-    flussConfig: Configuration)
-  extends FlussScan {
-
-  override protected val scanType: String = "Upsert"
+  override protected val scanType: String =
+    if (planner.hasLakeSnapshot) "LakeUpsert" else "Upsert"
 
   override def toBatch: Batch = {
     new FlussUpsertBatch(
       tablePath,
       tableInfo,
       readSchema,
-      partitionPredicate,
+      pushedPredicate,
       limit,
-      options,
-      flussConfig)
-  }
-
-  override def toMicroBatchStream(checkpointLocation: String): MicroBatchStream = {
-    new FlussUpsertMicroBatchStream(
-      tablePath,
-      tableInfo,
-      readSchema,
       options,
       flussConfig,
-      checkpointLocation)
-  }
-}
-
-/** Fluss Lake Upsert Scan for lake-enabled primary key tables. */
-case class FlussLakeUpsertScan(
-    tablePath: TablePath,
-    tableInfo: TableInfo,
-    requiredSchema: Option[StructType],
-    pushedPredicate: Option[FlussPredicate],
-    override val partitionPredicate: Option[FlussPredicate],
-    override val pushedSparkPredicates: Seq[Predicate],
-    override val limit: Option[Int],
-    options: CaseInsensitiveStringMap,
-    flussConfig: Configuration)
-  extends FlussScan {
-
-  override protected val scanType: String = "LakeUpsert"
-
-  override def toBatch: Batch = {
-    new FlussLakeUpsertBatch(
-      tablePath,
-      tableInfo,
-      readSchema,
-      pushedPredicate,
-      partitionPredicate,
-      limit,
-      options,
-      flussConfig)
+      planner)
   }
 
   override def toMicroBatchStream(checkpointLocation: String): MicroBatchStream = {
