+++ fluss-spark/fluss-spark-common/src/main/scala/org/apache/fluss/spark/read/lake/FlussLakePartitionReaderFactory.scala
@@ -42,7 +42,7 @@ class FlussLakePartitionReaderFactory(
   @transient private lazy val lakeSource: LakeSource[LakeSplit] = {
     val source = FlussLakeUtils.createLakeSource(flussConfig.toMap, tableProperties, tablePath)
     source.withProject(FlussLakeUtils.lakeProjection(projection))
-    flussPredicate.foreach(FlussLakeBatch.applyLakeFilters(source, _))
+    flussPredicate.foreach(FlussLakeUtils.applyLakeFilters(source, _))
     source
   }
